added changes for compiling in bazel-0.21.0

+++ heron/common/src/java/org/apache/heron/common/basics/ByteAmount.java
@@ -25,8 +25,8 @@ package org.apache.heron.common.basics;
 public final class ByteAmount implements Comparable<ByteAmount> {
   private static final long MB = 1024L * 1024;
   private static final long GB = MB * 1024;
-  private static final long MAX_MB = Math.round(Long.MAX_VALUE / MB);
-  private static final long MAX_GB = Math.round(Long.MAX_VALUE / GB);
+  private static final long MAX_MB = Math.round((double)Long.MAX_VALUE / MB);
+  private static final long MAX_GB = Math.round((double)Long.MAX_VALUE / GB);
 
   public static final ByteAmount ZERO = ByteAmount.fromBytes(0);
   private final long bytes;
